Check the return value from symbol_lookup.
1998-09-27  Miguel de Icaza  <miguel@nuclecu.unam.mx>

	* src/expr.c (function_call_with_values): Check the return value
	from symbol_lookup.

1998-09-26  Miguel de Icaza  <miguel@nuclecu.unam.mx>

	* src/sheet-object.c (sheet_view_object_realize): Swap coordinates
	before passing to the canvas item.

	* plugins/python/python.c (init_plugin): Use the system
	installation directory to load the python scripts.

	* autogen.sh: Hack to make libtool add library dependcy
	information (it modifies the libtool file after it has been
	created).

	* src/workbook.c (workbook_set_filename): New routine to set the
	filename being edited of the workbook.

	* src/sheet-object.c (create_object): Swap arguments to the
	ellipse object to guarrantee we pass valid information.

	* src/render-ascii.c (cell_region_render_ascii): Ok, remove the
	leaks now.

+++ src/cell.c
@@ -84,7 +84,7 @@ cell_set_halign (Cell *cell, StyleHAlignFlags halign)
 {
 	g_return_if_fail (cell != NULL);
 
-	if (cell->style->halign == halign)
+	if (cell->style->halign == (unsigned int) halign)
 		return;
 
 	cell_queue_redraw (cell);
